assertScenesEqual : Remove pathsToIgnore, add pathsToPrune argument
The difference in behaviour is that child paths of the pruned paths are also not considered.

There was previously an inconsistency in the definition between "Don't check what the children of this location are", and "but do visit the children".  This new version is more consistent.

The tests that use it have been updated to match - there were 3 that did rely on testing children of the ignored path, which have been fixed as follows:
* to test a scene which has stayed the same except for one object being deleted, compare to a DeleteObject instead of ignoring the objects location
* one test didn't need to pass pathsToIgnore at all ... it passes even when testing all locations?
* one test was testing everything but transform and bound everywhere, and testing transform and bound while ignoring all but one location.  Explicitly testing transform + bound at that one location is equivalent, and not significantly more verbose.

For that last case, if it comes up, it might be nice to be able to specify prefixes in both scenes, to be able to compare one branch.  This would just require a scenePlug1PathPrefix argument to match scenePlug2PathPrefix.  This wouldn't have the inconsistency issue because when testing a particular branch, we can still check the parent childNames to make sure that branch exists, we just don't test anything else in the scene above that point.