Context : Don't emit `changedSignal()` unnecessarily
I was a little reluctant to add the default constructor with the invalid state for `Value()`, since until now all `Values` have been guaranteed to be valid. But in the long run I think we will want it anyway, to represent variables that have been removed (without needing to shuffle `m_map` around).