Context : Replace `Ownership` with `CopyMode`
Attempting to emphasise that it only influences the copy operation and has no bearing on subsequent `set()` calls. Not sure how helpful this is, but the old `Copied` value was definitely misleading, as we now only actually make copies as a last resort, preferring to share ownership where we can.