StandardStyle : Handle alpha correctly when converting LinToSRGB
The previous code was creating matte edges on transparent images because it was applying a color space transform to premultiplied channels, and was also incorrectly applying a color space gamma to the alpha channel.

Note that to get perfectly correct results, we also need to handle unpremulting properly when we apply a colorspace on load ( see https://github.com/ImageEngine/cortex/pull/1167 ).  But even without that, this appears to yield a result that is slightly less wrong.