OpenImageIOReader : Support gaffer:integerMapping metadata
This is a string that can be set to:
"remap": ( the default ), which maps the whole range of integers to 0 - 1 ( this is OIIO's default behavour )
"reinterpret": preserve the exact bits of the integer as a float ( the only mapping that doesn't destroy data ). The resulting value doesn't really make sense as a float, it is just for use with tool that expect integers stored as floats ( like Gaffer's ID picking tools )
"cast": Use the closest float value. This is the clearest way to work with integers in Gaffer, but can only exactly represent integers that are small enough to have an exact floating point representation.